+++ site/models/funding.php
@@ -14,29 +14,7 @@ JLoader::register('CrowdfundingModelProject', CROWDFUNDING_PATH_COMPONENT_SITE .
 
 class CrowdfundingModelFunding extends CrowdfundingModelProject
 {
-    /**
-     * Method to get the profile form.
-     *
-     * The base form is loaded from XML and then an event is fired
-     * for users plugins to extend the form with extra fields.
-     *
-     * @param    array   $data     An optional array of data for the form to interrogate.
-     * @param    boolean $loadData True if the form is to load its own data (default case), false if not.
-     *
-     * @return   mixed   A JForm object on success, false on failure
-     * @since    1.6
-     */
-    public function getForm($data = array(), $loadData = true)
-    {
-        // Get the form.
-        $form = $this->loadForm($this->option . '.funding', 'funding', array('control' => 'jform', 'load_data' => $loadData));
-
-        if (!$form) {
-            return false;
-        }
-
-        return $form;
-    }
+    protected $formName = 'funding';
 
     /**
      * Method to get the data that should be injected in the form.